backend/requirements.txt
django==3.2
Pillow==9.5.0
python-dotenv==1.0.0
asgiref==3.6.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn
psycopg2-binary
PyJWT
pytz==2023.3
sqlparse==0.4.4
xhtml2pdf
drf-extra-fields==3.4.1
djoser==2.2.0
requests==2.30.0
requests-oauthlib==1.3.1
